Bug Fix and css style change
