--- src/components/SideNav/components/SideNavList.tsx
@@ -1,37 +1,0 @@
-import { FC } from 'react'
-
-const SideNavList: FC = () => {
-  const ListItems = [
-    {
-      id: 1,
-      item: 'My Reports',
-    },
-    {
-      id: 2,
-      item: 'Analytics',
-    },
-    {
-      id: 3,
-      item: 'Knowledge Base',
-    },
-    {
-      id: 4,
-      item: 'Help & Support',
-    },
-  ]
-
-  return (
-    <ul>
-      {ListItems.map(({ id, item }) => (
-        <li
-          className="cursor-pointer pl-[52px] px-auto border-l-[8px] border-background hover:bg-white  hover:bg-opacity-10  hover:border-white"
-          key={id}
-        >
-          {item}
-        </li>
-      ))}
-    </ul>
-  )
-}
-
-export default SideNavList
